Continute to change project.

+++ app/models/read_group.rb
@@ -1,5 +1,7 @@
 class ReadGroup < ActiveRecord::Base
   belongs_to :run
+  belongs_to :project
+
   has_many :gc_biases, inverse_of: :read_group, foreign_key: [:rg_id], dependent: :destroy
   has_many :insert_sizes, inverse_of: :read_group, foreign_key: [:rg_id], dependent: :destroy
   has_many :base_biases, inverse_of: :read_group, foreign_key: [:rg_id], dependent: :destroy
@@ -27,8 +29,19 @@ class ReadGroup < ActiveRecord::Base
   end
 
   def check_if_run_has_read_groups
-    if run.read_groups.count == 1
+    if run && run.read_groups.count == 1
       run.destroy
+    end
+  end
+
+  def self.update_project_name(old_name, new_name)
+    ReadGroup.where('project = ?', old_name).update_all(project: new_name)
+  end
+
+  def self.destroy_all_associations(params_id)
+    read_group = ReadGroup.find(params_id)
+    ReadGroup.reflect_on_all_associations(:has_many).map(&:name).each do |association|
+      read_group.send(association).delete_all
     end
   end
 end
